+++ docs/py/options/classes.py
@@ -41,3 +41,11 @@ itables.show(df, classes="display")
 
 # %% tags=["full-width"]
 itables.show(df, classes="display nowrap cell-border")
+
+# %% [markdown]
+# ```tip
+# You can change the default for all your notebooks and apps by creating an `itables.toml` file in the current or a parent directory, with e.g. this content:
+# ~~~
+# classes = ["display", "nowrap", "compact"]
+# ~~~
+# ```